consolidation of the UserSearchResult type definition
