+++ Web/src/features/whispers/components/RecipientSelector.tsx
@@ -6,7 +6,7 @@ import { Button } from '@/components/ui/button';
 import { Card, CardContent } from '@/components/ui/card';
 import { Input } from '@/components/ui/input';
 import { useUserSearch } from '@/features/users/hooks/useUserSearch';
-import { UserSearchResult } from '@/types/user';
+import { UserSearchResult } from '@/features/users/types';
 
 interface RecipientSelectorProps {
   selectedRecipient: UserSearchResult | null;
